OAK-11197: replaced sha-256 random value generation with murmur3

+++ oak-search-elastic/src/main/java/org/apache/jackrabbit/oak/plugins/index/elastic/query/async/facets/ElasticStatisticalFacetAsyncProvider.java
@@ -84,7 +84,7 @@ public class ElasticStatisticalFacetAsyncProvider implements ElasticFacetProvide
                         s.field(fs -> fs.field(
                                 ElasticIndexDefinition.PATH_RANDOM_VALUE)
                                 // this will handle the case when the field is not present in the index
-                                .unmappedType(FieldType.Short)
+                                .unmappedType(FieldType.Integer)
                         )
                 )
         );
